# server/openapi.yaml
openapi: 3.0.3
info:
  title: LEGION
  version: 0.1.0
  description: The engine. One Django server, every limb a client of it.
paths:
...
  /api/changes:
    get:
      operationId: api_changes_retrieve
      parameters:
      - in: query
        name: aspects
        schema:
          type: array
          items:
            type: string
            enum:
            - events
            - checklists
            - body
            - memory
            - places
            - voice_notes
        description: 'Comma-separated aspect names. Selects which top-level keys get
          populated: `checklists` fills `checklists`, `checklist_items` AND `checklist_ticks`;
          `body` fills all eight of its tables; `memory` all three. **Omitted or blank
          means every known aspect.** An unknown name is a 400 naming it - never a
          silently smaller response.'
        explode: false
        style: form
      - in: query
        name: since
        schema:
          type: string
          format: date-time
        description: ISO-8601 UTC watermark. Returns rows with `updated_at >= since`,
          tombstones included, oldest first. **Omitted or unparsable means EVERYTHING,
          never nothing** (api/sync.parse_since), so a fresh client's first pull is
          the whole table. Hand back the `next` cursor from the previous page, or
          `server_time` from GET /api/changes; both are already `Z`-suffixed so they
          need no extra encoding.
      tags:
      - changes
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Changes'
          description: 'One key per TABLE, named for the table, each holding every
            row changed at or after `since` with tombstones included, oldest first.
            **Not paged**: a large first pull should use the per-table `?since=` routes,
            which are.'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: An aspect name this server does not know. `detail` names it
            and lists the ones that exist.
...
components:
  schemas:
    BodyweightLog:
      type: object
      description: |-
        Base for every serializer this viewset drives.

        Two behaviours, both ticket 04's own text:

        1. **Unknown fields are refused, never dropped.** "A limb sending a
           column the server does not know is a version skew, and it says so
           with a 400 naming the field, never a silent drop." A phone one
           release ahead, writing a column this server has never heard of,
           must be told - a silent drop would look exactly like a successful
           write and lose the value.
        2. **The server facts are the server's.** `id`, `provenance`,
           `created_at`, `updated_at` and `deleted_at` are read-only on every
           subclass; `deleted_at` moves only via DELETE, `updated_at` only via
           the `touch_updated_at` trigger.

        Subclasses set `Meta.model` / `Meta.fields` and, where the table's
        `provenance` column is not the shared `USER` default,
        `default_provenance`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        weight_value:
          type: number
          format: double
        weight_unit:
          type: string
        logged_at:
          type: string
          format: date-time
        trust_tier:
          type: string
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
      required:
      - created_at
      - deleted_at
      - id
      - logged_at
      - origin_guid
      - provenance
      - trust_tier
      - updated_at
      - weight_unit
      - weight_value
    Changes:
      type: object
      properties:
        server_time:
          type: string
          format: date-time
          description: Postgres's own clock, read before any row query. Store THIS
            as the next `since`, rather than a max(updated_at) computed from the rows
            returned.
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        checklists:
          type: array
          items:
            $ref: '#/components/schemas/Checklist'
        checklist_items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        checklist_ticks:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistTick'
        places:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        voice_notes:
          type: array
          items:
            $ref: '#/components/schemas/VoiceNote'
        bodyweight_logs:
          type: array
          items:
            $ref: '#/components/schemas/BodyweightLog'
        meal_logs:
          type: array
          items:
            $ref: '#/components/schemas/MealLog'
        meal_targets:
          type: array
          items:
            $ref: '#/components/schemas/MealTarget'
        sleep_logs:
          type: array
          items:
            $ref: '#/components/schemas/SleepLog'
        sleep_targets:
          type: array
          items:
            $ref: '#/components/schemas/SleepTarget'
        workout_plans:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutPlan'
        workout_plan_items:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutPlanItem'
        workout_set_logs:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSetLog'
        memories:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        companion_memories:
          type: array
          items:
            $ref: '#/components/schemas/CompanionMemory'
        memory_audit:
          type: array
          items:
            $ref: '#/components/schemas/MemoryAudit'
      required:
      - server_time
    Checklist:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        schedule_kind:
          type: string
          nullable: true
        schedule_every:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        schedule_days_of_week:
          type: string
          nullable: true
        sort_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        archived:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        sync_id:
          type: string
          nullable: true
          maxLength: 64
      required:
      - created_at
      - deleted_at
      - id
      - name
      - updated_at
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        checklist:
          type: string
          format: uuid
        text:
          type: string
        sort_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        sync_id:
          type: string
          nullable: true
          maxLength: 64
        measure_unit:
          type: string
          nullable: true
          maxLength: 64
        measure_target:
          type: number
          format: double
          nullable: true
        measure_direction:
          type: string
          nullable: true
          maxLength: 16
      required:
      - checklist
      - created_at
      - deleted_at
      - id
      - text
      - updated_at
    ChecklistTick:
      type: object
      description: |-
        Read/representation shape only - a tick is never created or edited
        through this serializer directly; `ChecklistItemTickView` builds
        `ChecklistTick` rows itself so it can apply the revive-on-retick
        semantics `ChecklistController.tick`/`ChecklistTickDao.retick` describe,
        which a plain `ModelSerializer.create()` cannot express (it does not
        know to look for a soft-deleted row occupying the same `(item, day)`
        slot first).
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        item:
          type: string
          format: uuid
        day:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        ticked_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        sync_id:
          type: string
          nullable: true
          maxLength: 64
        value:
          type: number
          format: double
          nullable: true
        source:
          type: string
      required:
      - day
      - deleted_at
      - id
      - item
      - ticked_at
      - updated_at
...
    Detail:
      type: object
      description: |-
        `{"detail": "..."}` - the shape every hand-written refusal in this API
        uses, and the shape DRF itself uses for 401/403/404/405. The strings are
        written to be shown to a person: they say what did NOT happen (CLAUDE.md
        section 7's outcome-verb rule), so a client should surface `detail`
        rather than inventing its own wording.
      properties:
        detail:
          type: string
      required:
      - detail
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        all_day:
          type: boolean
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        source:
          type: string
        google_event_id:
          type: string
          nullable: true
        done:
          type: boolean
        done_at:
          type: string
          format: date-time
          nullable: true
        sort_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        trigger_place_label:
          type: string
          nullable: true
        repeat_kind:
          type: string
          nullable: true
        repeat_every:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_days_of_week:
          type: string
          nullable: true
        repeat_day:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_month:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        repeat_end_kind:
          type: string
          nullable: true
        repeat_end_date:
          type: string
          format: date
          nullable: true
        repeat_end_count:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        exact:
          type: boolean
        exact_downgraded:
          type: boolean
        missed_at:
          type: string
          format: date-time
          nullable: true
        missed_dismissed_at:
          type: string
          format: date-time
          nullable: true
        logged_at:
          type: string
          format: date-time
          nullable: true
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
          nullable: true
        structured_meta:
          nullable: true
        kind:
          type: string
      required:
      - created_at
      - deleted_at
      - id
      - provenance
      - title
      - updated_at
...
    MealLog:
      type: object
      description: |-
        Base for every serializer this viewset drives.

        Two behaviours, both ticket 04's own text:

        1. **Unknown fields are refused, never dropped.** "A limb sending a
           column the server does not know is a version skew, and it says so
           with a 400 naming the field, never a silent drop." A phone one
           release ahead, writing a column this server has never heard of,
           must be told - a silent drop would look exactly like a successful
           write and lose the value.
        2. **The server facts are the server's.** `id`, `provenance`,
           `created_at`, `updated_at` and `deleted_at` are read-only on every
           subclass; `deleted_at` moves only via DELETE, `updated_at` only via
           the `touch_updated_at` trigger.

        Subclasses set `Meta.model` / `Meta.fields` and, where the table's
        `provenance` column is not the shared `USER` default,
        `default_provenance`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        description:
          type: string
        calories_kcal:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
          description: 'An estimate, not a measurement: the model''s guess from the
            meal description. Never gated and never treated as a reconciled figure
            (CLAUDE.md section 4 rule 5).'
        protein_g:
          type: number
          format: double
          nullable: true
          description: 'An estimate, not a measurement: the model''s guess from the
            meal description. Never gated and never treated as a reconciled figure
            (CLAUDE.md section 4 rule 5).'
        carbs_g:
          type: number
          format: double
          nullable: true
          description: 'An estimate, not a measurement: the model''s guess from the
            meal description. Never gated and never treated as a reconciled figure
            (CLAUDE.md section 4 rule 5).'
        fat_g:
          type: number
          format: double
          nullable: true
          description: 'An estimate, not a measurement: the model''s guess from the
            meal description. Never gated and never treated as a reconciled figure
            (CLAUDE.md section 4 rule 5).'
        logged_at:
          type: string
          format: date-time
        source_image_path:
          type: string
          nullable: true
        trust_tier:
          type: string
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
      required:
      - created_at
      - deleted_at
      - description
      - id
      - logged_at
      - origin_guid
      - provenance
      - trust_tier
      - updated_at
    MealTarget:
      type: object
      description: |-
        Base for every serializer this viewset drives.

        Two behaviours, both ticket 04's own text:

        1. **Unknown fields are refused, never dropped.** "A limb sending a
           column the server does not know is a version skew, and it says so
           with a 400 naming the field, never a silent drop." A phone one
           release ahead, writing a column this server has never heard of,
           must be told - a silent drop would look exactly like a successful
           write and lose the value.
        2. **The server facts are the server's.** `id`, `provenance`,
           `created_at`, `updated_at` and `deleted_at` are read-only on every
           subclass; `deleted_at` moves only via DELETE, `updated_at` only via
           the `touch_updated_at` trigger.

        Subclasses set `Meta.model` / `Meta.fields` and, where the table's
        `provenance` column is not the shared `USER` default,
        `default_provenance`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        calories_kcal:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        protein_g:
          type: number
          format: double
        carbs_g:
          type: number
          format: double
        fat_g:
          type: number
          format: double
        effective_from_date:
          type: string
          format: date
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
      required:
      - calories_kcal
      - carbs_g
      - created_at
      - deleted_at
      - effective_from_date
      - fat_g
      - id
      - origin_guid
      - protein_g
      - provenance
      - updated_at
...
    Place:
      type: object
      description: |-
        Base for every serializer this viewset drives.

        Two behaviours, both ticket 04's own text:

        1. **Unknown fields are refused, never dropped.** "A limb sending a
           column the server does not know is a version skew, and it says so
           with a 400 naming the field, never a silent drop." A phone one
           release ahead, writing a column this server has never heard of,
           must be told - a silent drop would look exactly like a successful
           write and lose the value.
        2. **The server facts are the server's.** `id`, `provenance`,
           `created_at`, `updated_at` and `deleted_at` are read-only on every
           subclass; `deleted_at` moves only via DELETE, `updated_at` only via
           the `touch_updated_at` trigger.

        Subclasses set `Meta.model` / `Meta.fields` and, where the table's
        `provenance` column is not the shared `USER` default,
        `default_provenance`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        label:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - created_at
      - deleted_at
      - id
      - label
      - latitude
      - longitude
      - provenance
      - updated_at
...
    VoiceNote:
      type: object
      description: |-
        Field-for-field `RemoteVoiceNote`. See this module's doc comment for
        the audio column that is absent from both.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        transcript:
          type: string
          nullable: true
        kind:
          type: string
        provenance:
          allOf:
          - $ref: '#/components/schemas/VoiceNoteProvenanceEnum'
          readOnly: true
        interrupted:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - created_at
      - deleted_at
      - id
      - kind
      - provenance
      - started_at
      - updated_at
...
    WorkoutSetLog:
      type: object
      description: |-
        No `source_list_item_id`. `WorkoutSetLog.sourceListItemId` names a
        phone-local Room row id from a table that has no server counterpart, so
        it would be a dangling reference here - the body migration's own comment
        on that table, and `RemoteWorkoutSetLog`'s, both say so.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        exercise:
          type: string
        sets:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        reps:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
        weight_value:
          type: number
          format: double
          nullable: true
        weight_unit:
          type: string
          nullable: true
        logged_at:
          type: string
          format: date-time
        trust_tier:
          type: string
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
      required:
      - created_at
      - deleted_at
      - exercise
      - id
      - logged_at
      - origin_guid
      - provenance
      - sets
      - trust_tier
      - updated_at
  securitySchemes:
    deviceToken:
      type: apiKey
      in: header
      name: Authorization
      description: One token per device (phone, browser, future robot), revocable
        alone. Send the value as `Token <key>`, including the word Token and the space.
        Get a key from POST /api/auth/login; POST /api/auth/logout revokes only the
        key that made the request.